requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
decord==0.6.0
deepspeed==0.17.4
dill==0.3.8
easydict==1.13
einops==0.8.1
evaluate==0.4.5
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
hf-xet==1.1.5
hjson==3.1.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp310-cp310-linux_x86_64.whl#sha256=120568f3fd675f59e4cb6de79a5c193b4067d90878d8bab2040fda8e3a2df6fa
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.33.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rouge_score==0.1.2
safetensors==0.5.3
scikit-learn @ file:///croot/scikit-learn_1753427383173/work
scipy @ file:///croot/scipy_1747238027288/work/dist/scipy-1.15.3-cp310-cp310-linux_x86_64.whl#sha256=4055d6b52bbbfc19594ac7c25008889945de175e153afc0935b6415716fa348e
six==1.17.0
soupsieve==2.8
sympy==1.13.1
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
timm==1.0.19
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.55.4
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
